Errores de la tabla Turno corregidos

+++ src/main/java/com/upc/vitalco/repositorios/TurnoRepositorio.java
@@ -3,5 +3,5 @@ package com.upc.vitalco.repositorios;
 import com.upc.vitalco.entidades.Turno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TurnoRepositorio extends JpaRepository<Turno, Long> {
+public interface TurnoRepositorio extends JpaRepository<Turno, Integer> {
 }
